+++ doc/src/release_notes.rst
@@ -60,6 +60,8 @@ Common Changes
       by the consumer. This avoids a segfault seen in some circumstances.
     - Fixed bug when deciding Arrow datatype for numeric expressions
       (`issue 510 <https://github.com/oracle/python-oracledb/issues/510>`__)
+    - Fixed bug when fetching numeric data that has no decimal point but the
+      Arrow array has scale > 0
 
     Note the data frame support in python-oracledb 3.3 is a pre-release, and
     may change in a future version.